Add support for specifying kusk extension options at the path and http method level. (#50)
The idea behind this is to create an options chain where the lowest
level options take precedence over the higher level ones. E.g. an option
set at the path level will take precedence over the same option set at
the global level for that path. Likewise for the HTTP method level: If a
setting is present at the HTTP method level it overrides the same
setting set at its path level for that HTTP method.

If a setting isn't provided at the HTTP method or path level, then the
options "chain" should be traversed until an option is found, ideally at
the global level, and if not, then a sensible default value should be
chosen, usually disabling whatever feature the option relates to.