Ambassador - Provide possibility to disable everything at root level - with corresponding possibility to enable at path/method level  (#163)
* enable operation override when path is disabled. This allows for a blanket disable on all operations on path and then can expose operations when they are deemed ready by overriding the disabled field at the operation level

* Add global level disable option to enable blanket disable on all paths
and operations. Paths and Operations can then be enabled when they are
deemed to be ready by overriding the disabled field at the path and
field levels.

Change disabled field in SubOptions from boolean to pointer to a boolean
to implement the following semantics:
- If nil, not explicitely set, then check for value at level above -
 if at operation, check path level. If at path level, check global
level
- If has a value be it True or False, disregard the disabled setting at
  the level above as it has been explicitely set

Implement useful helper methods IsOperationDisabled and IsPathDisabled
with the Option struct reciever which implement the semantics listed
above

* Update generator to reflect the fact that the disabled field is now a
boolean pointer. Use Options helper methods for determining whether a
path or operation is disabled

* Update tests to reflect the fact that disabled is now a boolean pointer
for SubOptions

* Update traefik to use option helper methods for checking if path/operation is disabled